Move the registers in a dedicated directory
