--- packages/jupyter-chat/src/chat-commands/index.ts
@@ -1,7 +1,0 @@
-/*
- * Copyright (c) Jupyter Development Team.
- * Distributed under the terms of the Modified BSD License.
- */
-
-export * from './types';
-export * from './registry';
