+++ packages/jupyter-chat/src/index.ts
@@ -4,14 +4,12 @@
  */
 
 export * from './active-cell-manager';
-export * from './chat-commands';
 export * from './components';
-export * from './footers';
 export * from './icons';
 export * from './input-model';
 export * from './markdown-renderer';
 export * from './model';
-export * from './registry';
+export * from './registers';
 export * from './selection-watcher';
 export * from './types';
 export * from './widgets/chat-error';
