+++ packages/jupyter-chat/src/registers/footers.ts
@@ -4,7 +4,15 @@
  */
 
 import { Token } from '@lumino/coreutils';
-import { MessageFooter, MessageFooterSection } from './types';
+import { IChatModel } from '../model';
+import { IChatMessage } from '../types';
+
+/**
+ * The token providing the chat footer registry.
+ */
+export const IMessageFooterRegistry = new Token<IMessageFooterRegistry>(
+  '@jupyter/chat:ChatFooterRegistry'
+);
 
 /**
  * The interface of a registry to provide chat footer.
@@ -21,6 +29,32 @@ export interface IMessageFooterRegistry {
    */
   addSection(section: MessageFooterSection): void;
 }
+
+/**
+ * The props sent passed to each `MessageFooterSection` React component.
+ */
+export type MessageFooterSectionProps = {
+  model: IChatModel;
+  message: IChatMessage;
+};
+
+/**
+ * A message footer section which can be added to the footer registry.
+ */
+export type MessageFooterSection = {
+  component: React.FC<MessageFooterSectionProps>;
+  position: 'left' | 'center' | 'right';
+};
+
+/**
+ * The message footer returned by the registry, composed of 'left', 'center',
+ * and 'right' sections.
+ */
+export type MessageFooter = {
+  left?: MessageFooterSection;
+  center?: MessageFooterSection;
+  right?: MessageFooterSection;
+};
 
 /**
  * The default implementation of the message footer registry.
@@ -43,10 +77,3 @@ export class MessageFooterRegistry implements IMessageFooterRegistry {
 
   private _footers: MessageFooter = {};
 }
-
-/**
- * The token providing the chat footer registry.
- */
-export const IMessageFooterRegistry = new Token<IMessageFooterRegistry>(
-  '@jupyter/chat:ChatFooterRegistry'
-);
